added special case bools to showdata
is the "show" all ads, to run between the actual shows on that channel?
is the "show" a movie? either a single uninterrupted file or a multi-part file for ads or whatever? will be run without interruption (unless with ads)
does the show have specials? say a Halloween or Christmas themed episode or a multi-part episode like "who killed mister burns"